Update BDModulePilotAI.cs
Added Energy and Angles fighting modes, Turn Radius as a replacement for MinAltitude(currently serve as a hard deck), slightly different minSpeed reasoning depending on angles vs energy mode, Corner Speed, the speed the AI uses for evasive manuvers, and the min speed in energy mode, unless in steerMode == SteerModes.Aiming, ability to adjust the distance the AI reacts to missiles, and the ability to 1. toggle extensions on/off, and 2. if theyre on, a multiplier for the distance the AI extends.